+++ www/static/src/routes/product/FirmOrderTest.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+
+class FirmOrderTest extends Component {
+    constructor(){
+        super()
+    }
+
+    render(){
+        return(
+            <div>
+                test
+            </div>
+        )
+    }
+}
+
+export default FirmOrderTest;
